# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger UI for Cantor API",
        "title": "Canto API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "paths": {
        "/csr": {
            "get": {
                "description": "return json list of CSRs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Query CSR list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/csr/{id}": {
            "get": {
                "description": "return json object of CSR",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Query CSR by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CSR nft id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/dex/pairs": {
            "get": {
                "description": "return json array of all pairs in Canto dex",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Query all pairs in Canto dex",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/requestengine.Pairs"
                        }
                    }
                }
            }
        },
        "/gov/proposals": {
            "get": {
                "description": "return json list of proposals",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Query proposal list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/gov/proposals/{id}": {
            "get": {
                "description": "return json object of proposal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Query proposal by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "proposal id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/lending/ctokens": {
            "get": {
                "description": "return json array of all pairs in CLM",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Query all cTokens in CLM",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/staking/apr": {
            "get": {
                "description": "return string of current staking APR",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Query current staking APR",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/staking/validators": {
            "get": {
                "description": "return json list of validators",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Query validator list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/staking/validators/{address}": {
            "get": {
                "description": "return json object of validator",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Query validator by address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "validator address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "requestengine.Pairs": {
            "type": "object",
            "properties": {
                "blockNumber": {
                    "type": "string"
                },
                "pairsStruct": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requestengine.ProcessedPair"
                    }
                }
            }
        },
        "requestengine.ProcessedPair": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "atob": {
                    "type": "boolean"
                },
                "cdecimals": {
                    "type": "integer"
                },
                "clpaddress": {
                    "type": "string"
                },
                "decimals": {
                    "type": "integer"
                },
                "lpPrice": {
                    "type": "string"
                },
                "price1": {
                    "type": "string"
                },
                "price2": {
                    "type": "string"
                },
                "ratio": {
                    "type": "string"
                },
                "reserve1": {
                    "type": "string"
                },
                "reserve2": {
                    "type": "string"
                },
                "stable": {
                    "type": "boolean"
                },
                "symbol": {
                    "type": "string"
                },
                "token0": {
                    "$ref": "#/definitions/requestengine.Token"
                },
                "token1": {
                    "$ref": "#/definitions/requestengine.Token"
                },
                "totalSupply": {
                    "type": "string"
                },
                "tvl": {
                    "type": "string"
                }
            }
        },
        "requestengine.Token": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "chainId": {
                    "type": "integer"
                },
                "decimals": {
                    "type": "integer"
                },
                "logoURI": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        }
    }
}